refactor: fix imports in app layouts and not-found screen
